Casting result of processors_count instead of forcing type on second argument.

+++ extension/threadpool/threadpool.cpp
@@ -145,8 +145,8 @@ ThreadPool* get_threadpool() {
      * tricky to detect if we are running under tsan, for now capping the
      * default threadcount to the tsan limit unconditionally.
      */
-    constexpr long unsigned int tsan_thread_limit = 63;
-    return std::min(result, tsan_thread_limit);
+    constexpr unsigned int tsan_thread_limit = 63;
+    return std::min(static_cast<unsigned int>(result), tsan_thread_limit);
   })();
 
   static auto threadpool = std::make_unique<ThreadPool>(num_threads);
